[refactor] 패키지 분리 방법 수정
