+++ src/main/java/subway/line/controller/LineController.java
@@ -1,8 +1,10 @@
-package subway.line.presentation;
+package subway.line.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.line.dto.LineRequest;
+import subway.line.dto.LineResponse;
 import subway.line.service.LineService;
 
 import java.net.URI;
