+++ src/main/java/subway/station/controller/StationController.java
@@ -1,8 +1,10 @@
-package subway.station.presentation;
+package subway.station.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
 import subway.station.service.StationService;
 
 import java.net.URI;
