+++ src/main/java/subway/station/dao/StationDao.java
@@ -1,12 +1,12 @@
-package subway.station.domain;
+package subway.station.dao;
 
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import subway.section.domain.SectionDao;
 import subway.exceptions.BadRequestException;
+import subway.station.domain.Station;
 
 import java.util.List;
 import java.util.ArrayList;
