+++ src/main/java/subway/station/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.station.presentation;
+package subway.station.dto;
 
 public class StationRequest {
     private String name;
